numeric filtering refactor
